trivial: extract StateComputeResult to separate file
