+++ consensus/src/pipeline/tests/execution_phase_tests.rs
@@ -20,7 +20,7 @@ use aptos_consensus_types::{
     quorum_cert::QuorumCert,
 };
 use aptos_crypto::HashValue;
-use aptos_executor_types::{ExecutorError, StateComputeResult};
+use aptos_executor_types::{state_compute_result::StateComputeResult, ExecutorError};
 use aptos_types::{ledger_info::LedgerInfo, validator_verifier::random_validator_verifier};
 use async_trait::async_trait;
 use std::sync::{
