+++ execution/executor-test-helpers/src/lib.rs
@@ -10,7 +10,7 @@ use aptos_crypto::{
     HashValue,
 };
 use aptos_executor::db_bootstrapper::{generate_waypoint, maybe_bootstrap};
-use aptos_executor_types::StateComputeResult;
+use aptos_executor_types::state_compute_result::StateComputeResult;
 use aptos_storage_interface::DbReaderWriter;
 use aptos_types::{
     account_address::AccountAddress,
